Add support for MIP mapping and KTX texture format
- Fix a very nasty bug in ncSceneNode destructor
- Build Windows version out of a single include and lib folder
- Add private copy constructor to some classes
- Fix SDL and GLFW header inclusion on OS X

+++ include/ncThread.h
@@ -56,7 +56,7 @@ public:
 	void *Join();
 
 	// Returns the calling thread id
-	static unsigned int Self();
+	static long int Self();
 	// Terminates the calling thread
 	static void Exit(void *pRetVal);
 	// Yields the calling thread in favour of another one with the same priority
